Create checkConstraintfinal1.py
This file is used to check constraints for the game. There are three predict types, type 0 is used to check whether player's symbol is constantly placed in three rows or columns or in diagonal side. Type 1 is used to get the number of positions left for a player to win the game. Type 3 returns the blank  position from the board if the opponent player is going to win the game by one move and in type 3 direction is value is used to determine the opponent's winning strike position.